Moving FHIR Bundle handling to middleware
Previously, IndividualResult would take care of all of the heavy lifting of
parsing through the structure of FHIR bundles. This shouldn't be in the
component and could lead to the bundle being refiltered if the same result is
being viewed more than once. This cleans up the object structure before it gets
to the redux store.