Implementing the data set/matching system screen
The previous implementation was really focused on the latest run. This
restructures things so that the component that displays the information on a
data set/record matching system pair has a much better handle on what is going
on.

Changes:
* IndividualResult is gone. It has been replaced by RunList
* The app behaves better when selecting the matching system from the drop down
selector.
* Got rid of react-chartjs. It seems like it is more trouble than it is worth.
We can now use the latest version of chart.js
* Implemented the run history line chat. It is in dire need of styling.
* Switched JOB to RUN in types