Now selecting record sets and RMSs
Implemented via middleware. When match runs come in, the new middleware checks
to make sure that a context is selected. If so, it will go through the match
runs to find the appropriate record matching systems and record sets to select.

New action types and a reducer changes were made to accommodate this. Components
were also that their class names in the list groups appends action where needed.